+++ config/routes.rb
@@ -1,7 +1,6 @@
 Rails.application.routes.draw do
 
   if AuthRocket::Api.use_default_routes
-    get 'login' => 'logins#login'
     get 'logout' => 'logins#logout'
   end