+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - NOTE: This version includes breaking changes.
   It is only compatible with AuthRocket 2. Use gem version '~> 2.0' with AuthRocket 1.
+- Refactor Rails integration
 - Update resources for AuthRocket 2
   - Add: ClientApp, Connection, Domain, Invitation, NamedPermission, Oauth2Session, ResourceLink, Token
   - Remove: LoginPolicy, UserToken,